Create tarallo_interface.py
This file will contain a further abstraction of the pytarallo. It will
mainly contains methods previously present in turbofresa.py, in addition
to other pieces of code.
The reason behind this is to separate the main function of the
TURBOFRESA (cleaning disks) from the TARALLO connectivity features,
allowing the main program to be completely independent from the database
(in regard to #34 as well).