Calculate badblocks timeout dynamically
The timeout calculation for the badblocks process is calculated based on
the disk capacity. The `mins_per_gb` variable, as the name implies,
indicates roughly how many minutes it takes for badblocks to clean one
GiB of a disk. The value has been empirically calculated through testing
over a 32 GiB USB drive.